user added

+++ Office_Seat_Book/Office_Seat_Book_Entity/Employee.cs
@@ -18,7 +18,7 @@ namespace Office_Seat_Book_Entity
         public string Name { get; set; }
 
         [AllowNull]
-        public int PhoneNo { get; set; }
+        public double PhoneNo { get; set; }
 
         [Required]
         public string Email { get; set; }
